Check log4j2 log file when using Spark 3
When logsInDeveloperConsole(true) is called on NotebookSparkSessionBuilder,
it tries to find the location of the log4j file, watches it, and send its
content to the browser console.

This makes sure that the log4j2 file (rather than just log4j) can be found.
